SVM: Move RewardInfo from accounts-db to Solana SDK
