+++ accounts-db/src/rent_debits.rs
@@ -1,6 +1,5 @@
 use {
-    crate::stake_rewards::RewardInfo,
-    solana_sdk::{pubkey::Pubkey, reward_type::RewardType},
+    solana_sdk::{pubkey::Pubkey, reward_info::RewardInfo, reward_type::RewardType},
     std::collections::HashMap,
 };
