+++ geyser-plugin-manager/src/block_metadata_notifier.rs
@@ -4,13 +4,12 @@ use {
         geyser_plugin_manager::GeyserPluginManager,
     },
     log::*,
-    solana_accounts_db::stake_rewards::RewardInfo,
     solana_geyser_plugin_interface::geyser_plugin_interface::{
         ReplicaBlockInfoV3, ReplicaBlockInfoVersions,
     },
     solana_measure::measure::Measure,
     solana_metrics::*,
-    solana_sdk::{clock::UnixTimestamp, pubkey::Pubkey},
+    solana_sdk::{clock::UnixTimestamp, pubkey::Pubkey, reward_info::RewardInfo},
     solana_transaction_status::{Reward, Rewards},
     std::sync::{Arc, RwLock},
 };
