+++ geyser-plugin-manager/src/block_metadata_notifier_interface.rs
@@ -1,6 +1,5 @@
 use {
-    solana_accounts_db::stake_rewards::RewardInfo,
-    solana_sdk::{clock::UnixTimestamp, pubkey::Pubkey},
+    solana_sdk::{clock::UnixTimestamp, pubkey::Pubkey, reward_info::RewardInfo},
     std::sync::{Arc, RwLock},
 };
